passing data to child in loop
Place ur content in the new folder called Data in src and make a jsx file to store ur content. After that import that data in your app.js file where u want it to display and use the maping technic suppose blog is the object where data is stored. {blog.map((v,i)=>{return( <Productitems pitem={v} key={i}/> )})}. Pass it to its child/component and use that values in the component.